requirements.txt
numpy==1.12.1
pandas==0.19.2
Pillow==4.1.0
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
matplotlib==2.2.2
